requirements.txt
# Web3 & Blockchain
web3==6.15.1
eth-account==0.11.0

# Async Support
asyncio==3.4.3
aiohttp==3.9.3
aiofiles==23.2.1

# HTTP Requests
requests==2.31.0
httpx==0.26.0

# Data Processing
pandas==2.2.1
numpy==1.26.4

# API Framework
fastapi==0.110.0
uvicorn==0.27.1
flask==3.0.2

# Configuration & Validation
pydantic==2.6.3
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Notifications
python-telegram-bot==20.8

# Utilities
python-json-logger==2.0.7
tenacity==8.2.3
cachetools==5.3.3
colorlog==6.8.2

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-mock==3.12.0

# Type Hints
typing-extensions==4.10.0
